🐛 fixed the massMention bug for certain messages
If the message contained a special character following the mass mention (such as with `@everyone!`) it would bypass the regex filter and still trigger a Discord mention. By specifically specifying the explicit mass mention substring along with the `@`, this ensures it is captured in the conditonal no matter how it is written.